Optimized some php files.

+++ api/deleteAccount.php
@@ -7,11 +7,13 @@
 
     $id = "";
     $accountInfoData = [];
+    $fatalError = null;
 
     if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $id = getInput($_POST["id"]);
     }
 
+    // Get account info
     try {
         $stmt = $conn->prepare("SELECT * FROM accounts WHERE id = ?;");
         $stmt->bind_param("i", $id);
@@ -43,17 +45,17 @@
     try {
         // get followers id
         $sql = "SELECT followaccountid FROM follow WHERE followedaccountid=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
+        if (!$stmt) {
             echo json_encode(["success"=>false, "error"=>"follow_error"]);
             exit;
         }
 
-        $state->bind_param("i", $id);
+        $stmt->bind_param("i", $id);
 
-        $state->execute();
-        $result = $state->get_result();
+        $stmt->execute();
+        $result = $stmt->get_result();
 
         if ($result) {
             while ($row = $result->fetch_assoc()) {
@@ -73,73 +75,78 @@
                 $description = "We are sorry to inform you that " . $accountInfoData['username'] . " has deleted his account. You see this notification because you were following it. You will no longer receive notifications about activity on this account.";
 
                 $stmt->bind_param("isssi", $idNot, $titleNot, $description, $today, $zero);
-
                 $stmt->execute();
-                
                 $stmt->close();
             }
         }
         
-        $state->close();
+        $stmt->close();
 
         // follow clear
         $sql = "DELETE FROM follow WHERE followaccountid=? OR followedaccountid=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
-            echo json_encode(["success"=>false, "error"=>"follow_error"]);
+        if (!$stmt) {
+            echo json_encode(["success"=>false, "error"=>"delete_follow_error"]);
             exit;
         }
 
-        $state->bind_param("ii", $id, $id);
-
-        $state->execute();
-        
-        $state->close();
+        $stmt->bind_param("ii", $id, $id);
+        $stmt->execute();
+        $stmt->close();
 
         // accountideadata clear
         $sql = "DELETE FROM accountideadata WHERE accountid=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
-            echo json_encode(["success"=>false, "error"=>"accountideadata_error"]);
+        if (!$stmt) {
+            echo json_encode(["success"=>false, "error"=>"delete_accountideadata_error"]);
             exit;
         }
 
-        $state->bind_param("i", $id);
-
-        $state->execute();
-        
-        $state->close();
+        $stmt->bind_param("i", $id);
+        $stmt->execute();
+        $stmt->close();
 
         // reports clear
         $sql = "DELETE FROM reports WHERE accountid=? OR authorid=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
-            echo json_encode(["success"=>false, "error"=>"reports_error"]);
+        if (!$stmt) {
+            echo json_encode(["success"=>false, "error"=>"delete_reports_error"]);
             exit;
         }
 
-        $state->bind_param("ii", $id, $id);
+        $stmt->bind_param("ii", $id, $id);
+        $stmt->execute();
+        $stmt->close();
 
-        $state->execute();
-        
-        $state->close();
+        // notifications clear
+        $sql = "DELETE FROM notifications WHERE accountid=?;";
+        $stmt = $conn->prepare($sql);
+
+        if (!$stmt) {
+            echo json_encode(["success"=>false, "error"=>"delete_notifications_error"]);
+            exit;
+        }
+
+        $stmt->bind_param("i", $id);
+        $stmt->execute();
+        $stmt->close();
 
         // get ideas id
         $sql = "SELECT id FROM ideas WHERE authorid=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
+        if (!$stmt) {
             echo json_encode(["success"=>false, "error"=>"get_ideas_ids_error"]);
             exit;
         }
 
-        $state->bind_param("i", $id);
+        $stmt->bind_param("i", $id);
 
-        $state->execute();
-        $result = $state->get_result();
+        $stmt->execute();
+        $result = $stmt->get_result();
 
         $ideasId = [];
 
@@ -149,82 +156,61 @@
             }
         }
         
-        $state->close();
-
-        // notifications clear
-        $sql = "DELETE FROM notifications WHERE accountid=?;";
-        $state = $conn->prepare($sql);
-
-        if (!$state) {
-            echo json_encode(["success"=>false, "error"=>"delete_notifications_error"]);
-            exit;
-        }
-
-        $state->bind_param("i", $id);
-
-        $state->execute();
-        
-        $state->close();
+        $stmt->close();
 
         for ($i=0; $i < count($ideasId); $i++) { 
             // idealabels clear
             $sql = "DELETE FROM idealabels WHERE ideaid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"idealabels_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
 
             // reports clear
             $sql = "DELETE FROM reports WHERE ideaid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"reports_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
 
             // follow clear
             $sql = "DELETE FROM follow WHERE followedideaid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"follow_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
 
             // comments clear
             $sql = "SELECT id FROM comments WHERE ideaid=? AND superCommentid IS NULL;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"comments_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
+            $stmt->bind_param("i", $ideasId[$i]);
 
-            $state->execute();
-            $result = $state->get_result();
+            $stmt->execute();
+            $result = $stmt->get_result();
 
             if ($result) {
                 while ($row = $result->fetch_assoc()) {
@@ -232,70 +218,67 @@
                 }
             }
             
-            $state->close();
+            $stmt->close();
+
+            if ($fatalError != null) {
+                echo json_encode(["success"=>false, "error"=>$fatalError]);
+                exit;
+            }
 
             // authorupdates clear
             $sql = "DELETE FROM authorupdates WHERE ideaid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"authorupdates_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
 
             // additionalinfo clear
             $sql = "DELETE FROM additionalinfo WHERE ideaid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"additionalinfo_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
 
             // ideas clear
             $sql = "DELETE FROM ideas WHERE id=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"ideas_error"]);
                 exit;
             }
 
-            $state->bind_param("i", $ideasId[$i]);
-
-            $state->execute();
-            
-            $state->close();
+            $stmt->bind_param("i", $ideasId[$i]);
+            $stmt->execute();
+            $stmt->close();
         }
 
         // accounts clear
         $sql = "DELETE FROM accounts WHERE id=?;";
-        $state = $conn->prepare($sql);
+        $stmt = $conn->prepare($sql);
 
-        if (!$state) {
+        if (!$stmt) {
             echo json_encode(["success"=>false, "error"=>"accounts_error"]);
             exit;
         }
 
-        $state->bind_param("i", $id);
-
-        $state->execute();
-        
-        $state->close();
+        $stmt->bind_param("i", $id);
+        $stmt->execute();
+        $stmt->close();
     } catch (\Throwable $th) {
-        echo json_encode(["success"=>false, "error"=>$th->__toString()]);
+        echo json_encode(["success"=>false, "error"=>strval($th)]);
         exit;
     }
 
@@ -316,21 +299,21 @@
     function deleteAllIdsSubComments($conn, $id) {
         try {
             $sql = "SELECT * FROM comments WHERE superCommentid=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"database_connection"]);
                 exit;
             }
 
-            $state->bind_param("i", $id);
+            $stmt->bind_param("i", $id);
             
-            if (!$state->execute()) {
+            if (!$stmt->execute()) {
                 echo json_encode(["success"=>false, "error"=>"execution_command_delete_subcomments"]);
                 exit;
             }
 
-            $result = $state->get_result();
+            $result = $stmt->get_result();
 
             $data = [];
 
@@ -338,7 +321,7 @@
                 $data[] = $row;
             }
 
-            $state->close();
+            $stmt->close();
 
             // $data contains all the subcomments
             foreach ($data as $row) {
@@ -348,24 +331,24 @@
             // This is the deeper ($id comment)
             // Delete the subcomment
             $sql = "DELETE FROM comments WHERE id=?;";
-            $state = $conn->prepare($sql);
+            $stmt = $conn->prepare($sql);
 
-            if (!$state) {
+            if (!$stmt) {
                 echo json_encode(["success"=>false, "error"=>"database_connection"]);
                 exit;
             }
 
-            $state->bind_param("i", $id);
+            $stmt->bind_param("i", $id);
             
-            if (!$state->execute()) {
+            if (!$stmt->execute()) {
                 echo json_encode(["success"=>false, "error"=>"execution_command_delete_subcomments"]);
                 exit;
             }
 
-            $state->close();
+            $stmt->close();
         } catch (\Throwable $th) {
             global $fatalError;
-            $fatalError = $th;
+            $fatalError = strval($th);
         }
         
         return;
